BVTCK-189 Mark the JavaFX tests with the newly introduced annotation
